move Message object to common types
