+++ data/slack-jackson-dto/src/main/kotlin/com/kreait/slack/api/contract/jackson/group/chat/PostMessage.kt
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.annotation.JsonSubTypes
 import com.fasterxml.jackson.annotation.JsonTypeInfo
 import com.kreait.slack.api.contract.jackson.common.messaging.Attachment
 import com.kreait.slack.api.contract.jackson.common.messaging.Block
+import com.kreait.slack.api.contract.jackson.common.types.Message
 import com.kreait.slack.api.contract.jackson.util.InstantToString
 import com.kreait.slack.api.contract.jackson.util.JacksonDataClass
 import java.time.Instant
@@ -33,19 +34,6 @@ data class SuccessfulPostMessageResponse constructor(override val ok: Boolean,
                                                      @JsonProperty("channel") val channel: String,
                                                      @InstantToString @JsonProperty("ts") val timestamp: Instant,
                                                      @JsonProperty("message") val message: Message? = null) : PostMessageResponse(ok) {
-    companion object
-}
-
-@JacksonDataClass
-data class Message(
-        @JsonProperty("type") val type: String? = null,
-        @JsonProperty("subtype") val subtype: String? = null,
-        @JsonProperty("text") val text: String? = null,
-        @InstantToString @JsonProperty("ts") val timestamp: Instant,
-        @JsonProperty("username") val username: String? = null,
-        @JsonProperty("attachments") val attachments: List<Attachment>? = listOf(),
-        @JsonProperty("blocks") val blocks: List<Block>? = listOf(),
-        @JsonProperty("bot_id") val botId: String? = null) {
     companion object
 }
